idearadar/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.4
feedparser==6.0.11
requests==2.31.0
beautifulsoup4==4.12.3
trafilatura==1.8.0
simhash==2.1.2
python-dateutil==2.8.2
robotexclusionrulesparser==1.7.1
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
